# documentation/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalyzeRequest": {
        "description": "An analytics job on a projection of a namespace.",
        "properties": {
          "job": {
            "$ref": "#/components/schemas/Job"
          },
          "namespace": {
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/QueryOptions"
          },
          "projection": {
            "$ref": "#/components/schemas/Projection"
          }
        },
        "type": "object"
      },
      "AnalyzeResponse": {
        "description": "One chunk of a job's result: rows of the kind `kind` names (the same in\nevery chunk).",
        "properties": {
          "counts": {
            "items": {
              "$ref": "#/components/schemas/Count"
            },
            "type": "array"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/Group"
            },
            "type": "array"
          },
          "kind": {
            "$ref": "#/components/schemas/JobResultKind"
          },
          "meta": {
            "$ref": "#/components/schemas/AnswerMeta"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/Score"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "Error": {
        "description": "An error as a message, for access methods that carry errors in a body\n(REST, step 12): the same code and message as the gRPC status.",
        "properties": {
          "code": {
            "description": "One of the codes above.",
            "type": "string"
          },
          "message": {
            "description": "For people.",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ExplainRequest": {
        "description": "How `Find` would read a filter.",
        "properties": {
          "analyze": {
            "description": "Also run the index lookup and report the exact number of candidates.",
            "type": "boolean"
          },
          "filter": {
            "$ref": "#/components/schemas/Expr"
          },
          "namespace": {
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/QueryOptions"
          }
        },
        "type": "object"
      },
      "ExplainResponse": {
        "properties": {
          "explain": {
            "$ref": "#/components/schemas/Explain"
          },
          "meta": {
            "$ref": "#/components/schemas/AnswerMeta"
          }
        },
        "type": "object"
      },
...
      "NeighbourhoodRequest": {
        "description": "The nodes within `depth` edges of the seeds; paginated by id.",
        "properties": {
          "depth": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Uint64"
              }
            ],
            "description": "How many edges away a node may be (0: the seeds only)."
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          },
          "edgeFilter": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Expr"
              }
            ],
            "description": "Follow only edges matching this filter."
          },
          "edgeTypes": {
            "description": "Follow only edges with one of these types (all if empty).",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "namespace": {
            "type": "string"
          },
          "nodeFilter": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Expr"
              }
            ],
            "description": "Return only nodes matching this filter (it doesn't prune the search)."
          },
          "options": {
            "$ref": "#/components/schemas/QueryOptions"
          },
          "seeds": {
            "description": "Ids that don't exist are skipped.",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "NeighbourhoodResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/AnswerMeta"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/Node"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "SubgraphRequest": {
        "description": "The nodes within `depth` edges of the seeds and the edges between them.",
        "properties": {
          "depth": {
            "$ref": "#/components/schemas/Uint64"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          },
          "edgeFilter": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Expr"
              }
            ],
            "description": "Follow, and return, only edges matching this filter."
          },
          "edgeTypes": {
            "description": "Follow, and return, only edges with one of these types (all if empty).",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "namespace": {
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/QueryOptions"
          },
          "seeds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "SubgraphResponse": {
        "properties": {
          "edges": {
            "description": "Sorted by id over the whole stream.",
            "items": {
              "$ref": "#/components/schemas/Edge"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/AnswerMeta"
          },
          "nodes": {
            "description": "Sorted by id over the whole stream.",
            "items": {
              "$ref": "#/components/schemas/Node"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "description": "A session token from `POST /v1/auth/login`, or an API token.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "The console's session (login with `cookie: true`). Requests other than GET and HEAD must also send the `X-Iwdb-Csrf` header.",
        "in": "cookie",
        "name": "iwdb_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The `Database` trait over HTTP/JSON, with the messages of the gRPC contract (`proto/ironweaver_db/v1`) in their proto3 JSON form. See documentation/api/rest.md.",
    "license": {
      "identifier": "AGPL-3.0-only",
      "name": "AGPL-3.0-only (or a commercial license)"
    },
    "title": "Ironweaver DB REST API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/namespaces/{ns}/analyze": {
      "post": {
        "description": "An analytics job on a projection of a namespace.",
        "operationId": "analyze",
        "parameters": [
          {
            "description": "The namespace.",
            "in": "path",
            "name": "ns",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              },
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            },
            "description": "The answer. With `Accept: application/x-ndjson`, the answer's chunks, one per line, `meta` in the last; otherwise one message."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A request the server refuses: the HTTP status of its code (documentation/api/errors.md)."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A failure of the server: the HTTP status of its code (documentation/api/errors.md)."
          }
        },
        "summary": "Run an analytics job"
      }
    },
...
    "/v1/namespaces/{ns}/explain": {
      "post": {
        "description": "How `Find` would read a filter.",
        "operationId": "explain",
        "parameters": [
          {
            "description": "The namespace.",
            "in": "path",
            "name": "ns",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainResponse"
                }
              }
            },
            "description": "The answer."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A request the server refuses: the HTTP status of its code (documentation/api/errors.md)."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A failure of the server: the HTTP status of its code (documentation/api/errors.md)."
          }
        },
        "summary": "How find would read a filter"
      }
    },
...
    "/v1/namespaces/{ns}/neighbourhood": {
      "post": {
        "description": "The nodes within `depth` edges of the seeds; paginated by id.",
        "operationId": "neighbourhood",
        "parameters": [
          {
            "description": "The namespace.",
            "in": "path",
            "name": "ns",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NeighbourhoodRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighbourhoodResponse"
                }
              },
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/NeighbourhoodResponse"
                }
              }
            },
            "description": "The answer. With `Accept: application/x-ndjson`, the answer's chunks, one per line, `meta` in the last; otherwise one message."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A request the server refuses: the HTTP status of its code (documentation/api/errors.md)."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A failure of the server: the HTTP status of its code (documentation/api/errors.md)."
          }
        },
        "summary": "Nodes near seeds"
      }
    },
...
    "/v1/namespaces/{ns}/subgraph": {
      "post": {
        "description": "The nodes within `depth` edges of the seeds and the edges between them.",
        "operationId": "subgraph",
        "parameters": [
          {
            "description": "The namespace.",
            "in": "path",
            "name": "ns",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubgraphRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubgraphResponse"
                }
              },
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/SubgraphResponse"
                }
              }
            },
            "description": "The answer. With `Accept: application/x-ndjson`, the answer's chunks, one per line, `meta` in the last; otherwise one message."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A request the server refuses: the HTTP status of its code (documentation/api/errors.md)."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A failure of the server: the HTTP status of its code (documentation/api/errors.md)."
          }
        },
        "summary": "An induced subgraph"
      }
    },
    "/v1/namespaces/{ns}/traverse": {
      "post": {
        "description": "A breadth- or depth-first traversal, in traversal order.",
        "operationId": "traverse",
        "parameters": [
          {
            "description": "The namespace.",
            "in": "path",
            "name": "ns",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraverseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraverseResponse"
                }
              },
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/TraverseResponse"
                }
              }
            },
            "description": "The answer. With `Accept: application/x-ndjson`, the answer's chunks, one per line, `meta` in the last; otherwise one message."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A request the server refuses: the HTTP status of its code (documentation/api/errors.md)."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A failure of the server: the HTTP status of its code (documentation/api/errors.md)."
          }
        },
        "summary": "Breadth or depth first order"
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    },
    {
      "sessionCookie": []
    }
  ],
  "servers": [
    {
      "url": "http://127.0.0.1:7600"
    }
  ]
}